BUG: Using a pointer to data that goes out of scope
itkFEMElement2DTest was failing on gcc 4.7 platforms. The test was
assigning a pointer to data that went out of scope before the pointer
was used. This data was the expected solutions for a number of test
data sets.

The patch moves the expected solutions data declarations into the main
scope.

Change-Id: If229440a6a2e07309910cbc5621a97edaa1e8212