BUG: GPU tests sometimes fails
GPUGradientAnisotropicDiffusionImageFilterTest2D was failing on a
certain Mac OS X 10.7.2 and 10.7.4 configuration even though the 3D
version passed there and the 2D version passed on the same hardware
with OS X 10.6.8.  After much effort, it was determined that this
patch fixes the problem.  It had been documented that Apple OpenCL
implementations weren't handling nested if statements like if(A && B
&& C), but in this case, the implementation wasn't working for if (A
&& B).  Breaking into two individual statements seems to work, but it
is not clear whether other compound if statements are affected.  Right
now, they seem to be working.

However, this compound if statement could be the issue for
GPUMeanImageFilterTest3D which is failing on other Mac
configurations.  Therefore, this patch cleans up those compound if
statements as well.  The GPUMeanImageFilterTest has also been modified
to output the result images in case this patch does not fix the problem.

Change-Id: Ia13343bc557b77b53abfb3da4a091d11d87a34a1