COMP: Fix definition of Rigid3DTransform when LEGACY enabled
Thanks WrapITK for actually trying to instantiate all methods!

Previously the BackTransform's definition was dependent on ITK v3
legacy being defined, where the declaration was dependent only the
ITK_LEGACY_REMOVE not being defined.

The logic was cleaned up in the hxx file, and comments added to ending
pre-processor conditionals.

Change-Id: I4a6785b48426c905f18ffe8c3bdf051e05e2dfb8