BUG: Use the input type for the threshold value
The input type is compared to the input pixel, it must be of the input
pixel type to be able to threshold the entire range. Previously,
trucation was occurring when the filter's ivar was being set.

Additionally, the pixel type of the foreground value was corrected in
certain places.

Documentation corrected.

These errors were detected during SimpleITK wrapping and testing.

Change-Id: Icba802f5b0b2a92985104da40331b0a80f0bb50e