PERF: Pull expensive GetInput/GetOutput calls out of inner loops
Between ITK3 and ITK4 a new scheme was set up for managing filter
inputs and outputs by name instead of by index.  This has made
the GetInput & GetOutput methods much slower. This has had a
serious performance penalty in several filters who call these
functions in inner loops.  In some cases this is a
straightforward fix, in others the input and output handles need
to be cached at the class level.

The fixes in this patch were accomplished by instrumenting the
ProcessObject methods and throwing an exception if they were
called an unreasonable number of times.  I started with a
threshold of 100,000 and after an initial pass, decreased this to
2000, at which point 'false positives' showed up.

Outstanding Issue:

PipelineMonitorImageFilter::PropogateRequestedRegion
Test Failure: itkJoinSeriesImageFilterStreamingTest
Solution: Cache pointers to input and output
Problem: Not sure where these could be initialized in such a way
that they would be updated if the input or output changed.

BinaryFunctorImageFilter::ThreadedGenerateData
Test Failure:  itkLabelMapContourOverlayImageFilterTest2
Solution: Cache input/output, possibly in BeforeThreadedGenerateData
Problem: I suspect that the test failure arises more from the
re-use of an instance of SubtractImageFilter, than any actual
problem of over-use.

WarpVectorImageFilter::GenerateInputRequestedRegion
Test Failures: itkDiffemorphicDemonsRegistrationFilterTests
(those numbered 01,02,03,04,05,06,07,09,10,11)
Solution: Not clear if one is needed; or if a simple fix is
possible.
Problem: itkExponentialDisplacementFieldImageFilter::GenerateData
uses a sort of a mini-pipeline; in a loop it re-uses the same
instance of WarpVectorImageFilter.  This racks up the count of
calls to GetInput/GetOutput in
WarpVectorImageFilter::GenerateInputRequestedRegion. It might be
good to go back and re-examine how this filter is built.

Change-Id: I2a92882098be6e8d026ff92d3de057f049b55cc7