PERF: Faster distance computation and fixed reflective iterator bug
Sped up the distance map computation by ignoring non-zero values in
the original image. The goal of the distance map computation by
definition is to find the distance at every pixel that has a value of
zero to its closest non-zero neighbor.  Therefore, the non-zero
neighbor values can be ignored when solving for the distance map, but
the old version simply processed all pixels in the image.  When large
portions of the image are non-zero, this results in significant
overhead.  The new version simply ignores these non-zero values from
the original image.  Added a test to
itkDanielssonDistanceMapImageFilterTest.cxx that creates a 3D image
that is mostly non-zero.  The distance map computation before and
after this fix is the same, but the new version is about 30 times
faster.

Corrected a bug in the ReflectiveImageRegionIterator that resulted in
differences in the pixel value returned by an iterator and the pixel
value defined by image at the iterator index.  This is a bizarre bug
because calling "rit.Get()" gives a different result than
"myImage->GetPixel( rit.GetIndex())", where "rit" is the iterator
iterating over "myImage".  I added a test in
itkReflectiveImageRegionIteratorTest.cxx that checks these values
across the image, and this test fails.  However, with the new fix, it
passes, as expected.  The problem is that the m_PositionIndex is
incremented and then tested, but the m_Position is only incremented if
the test passes.  Thus, they become out of sync.  Now, the test checks
whether one past the current m_PositionIndex and then increments both
if the test passes.  Also, previously m_Position was not reset when
the iterator reached the end of a line even though m_PositionIndex was
reset; now m_Position is also reset.

Change-Id: I224ebc5b0814c17789be4fabf58a3280dc4bbb0c