COMP: Revert use of itkLegacyMacro
The itkRigid3DTransform class was intended
to be an abstract base class for Rigid Transform
child classes that adds constraints to the SetMatrix
funciton.

In ITKv3 this class was often mis-used.  In our attempts
to maintain backwards compatibility/warn on use of legacy
code, and provide a forward clean path for ITKv4 this
base class has become very difficult to set the proper usage
on member functions that is compatible with all the desired
compilation options.

When ITKv3 compatibility is selected, the GetInverse and
GetInverseTransform functions must be traversed from
child classes, and are therefore part of protected interface
that can not be designated as itkLegacyMacro (otherwise warnings
are issued even though they are not called directly).

In ITKv4, they do not need to be defined at all because
they can simply use the MatrixOffsetBase class version
of those functions.

Change-Id: Id0651d1ecc5ada38470f139325883065e5e2cc53