PERF: add alignment data attribute to per-thread structure
To reduce false sharing between thread, it can be benefitial to align
data to cachelines, to enforce padding of data structures.

Adding macro itkAlignedTypedef, which adds an alignment specifier to
an existin type to define a new type. This is made mostly portable
with the addition of the up and coming C++11 alignas type specifier,
and support for the native MSC and GNU extensions for alignment
specification. A try compile is added to detect the C++11 alignas
feature as well as the GNU attribute extension.

Change-Id: I97ab01edfb30ce43ea7d13fbbcaa58bb5d43f30a