PERF: remove per sample allocation, use per-thread structure
This patch addresses performacne regressions with ITKv3 registration
metrics and is intended for the release branch.

First there was a dynamic allocation that was going on a per
smaple for many metrics, this was for the jacobian. This was extracted
into a per-thread data structure or a member variable.

Additionally, there was a per-thread array for several data
types. This was changed to an per-thread structure to reduce false
sharing. Ideally, this structure should be aligned to the cache block
size, but this is awaiting a portable solution.

Change-Id: I57099f5bb649ba5db51aee8b0fc1f7ef1fca075a