BUG: Fix WrapITK GCCXML ExternalProject dependency.
Redundant dependencies specified in the wrong location were causing multiple
simultaneous attempts at fetching GCCXML, which would cause CVS to fail.  The
extra dependency is removed.  Note that the dependency is already specified in
the itk_end_wrap_submodule_gccxml macro.

Also the most appropriate dependency for completion of an ExternalProject is
<ProjectName> instead of <ProjectName>-install.  This fixes build errors where
the build of GCC_XML complete, but make fails, requiring a manual restart of
make.

Change-Id: I2b289e1e55bacb11b84e0b62f0c2259cd7a5e99e