BUG: Fix segfault from integer overflow in NIFTI reader
The type unsigned int was used for the size of an image this could
overflown for large images and cause a segfault. Specifically, this is
for vector images.  This type has been changed to size_t to ensure
that the in memory image's size can be computed.

A test was added to write a VectorImage of 3 components for a 6GB
file. This test is conditional on the computer having 12GB of memory.

Fix Segfault with Redhat's gcc 4.4.6, in NiftiImageIOTest, due to some
type of optimization bug in MakeNifiImage utility method.

Change-Id: I3e2c9d2f2c134ea1a379f188155763fed54bab1d