PERF: Avoid memcpy call per component for Nift
A function call to memcpy was done on a per component basic for
certain vector image in the nifti ImageIO. This has been changed to a
simple for loop. This enables the compiler to optimize the nexted loop
and perhaps perform vectorization of the code, in addition to removing the
expence of a function call to copy a single byte.

For an vector image of unsigned chars ~50% speedup was profiled,
with SSD, while for a vector image of float ~20% speedup was
profiled.

Change-Id: I2ce83065e15fd87622d10eafb7bc3518e10f5cb3