WIP: Force explicit instantiation/optimization
In order to profile external applicstion built upon
itk efficiently, ITK core libary should be built
in an optimized mode, and only your application
components that need debugging should be built
with debug instrumentation.

It was found that the itkIndex was the largest
peformance bottle neck because it was only defined
in the header, and therefore it was not possible
to separate the optimization of the itkIndex class
from the algorithm that needed debugging.

By putting the itkIndex<3u> implementation in a
highly optimized separate compilation unit, the
real performance issues of the external application
becomes much easier to debug.

In order to get extern templates to work it is
REQUIRED that the definition of the member functions
occurs outside of the class.  Any definition of
member functions that occur inside the class are
always assumed to be inline and are therefore
ignored by the "extern template", thus they are
compiled into every compilation unit ignoring the
benefits of separate one-time comiplation.

Change-Id: Ia806af7a0deb183a1b292f08e107433a16494bbf