ENH: added suport for buffered region in GPU image
1. In GPUImageDataManager, two arrays for the buffered region are added:
m_BufferedRegionIndex and m_BufferedRegionSize, for the starting index
and size of the buffered region respectively.

2. In GPUKernelManager, a new method SetKernelArgWithImageAndBufferedRegion
is added to pass to the GPU kernel the image buffer together with the
the starting index and size of the GPU buffered region.

3. In an OpenCl program -- GPUNeighborhoodOperatorImageFilter.cl, the code
is changed to reflect the buffered region information. The main change
is about the indices in the input image:

inputBufferIndex = outputBufferIndex + outStartIndex - inStartIndex

where outputBufferIndex is the relative index in the output buffered
region, outStartIndex is the starting index of that region, and similar
for inputBufferIndex and inStartIndex.

4. In the calling program, GPUNeighborhoodOperatorImageFilter, the passing
of parameters is simplified by automatically passing buffered region
information.

Change-Id: Ic56e35ded7ea51b39bee650070757e42a818e6d0