ENH: Add tests for iterator indices
Enforce index to be the same on the input and output for the PointsContainer,
CellsContainer, PointDataContainer, CellDataContainer.

The tests now run for itk::Mesh and itk::QuadEdgeMesh even if one of them is
failing. This information is quite useful to track down errors on the
dashboard. Add some printouts to know for which type the test is failing.

Change-Id: I37dce8019ac53f71a66a9387fae0f6437f1ef780