BUG: Removed too stringent ConceptChecking
The LabelImageGaussianInterpolateImageFunction is most natural
for integer type images, but this is not strictly a requirement.

The filter works find for floating point images if the number
of unique floating point values is small.  It is often the
case that this filter will be used on conceptually integer
data that is stored in floating point images for convenience.

Change-Id: Idbac400ad0e4de9bff505b20fc10f092bb608f21