BUG: SymmetricEigenAnalysis OrderByMagnitude bug
As reported in
http://old.nabble.com/Bug-in-SymmetricEigenAnalysis--p10157786.html
and patched as described in the Jira ITK-2939 issue here
(https://itk.icts.uiowa.edu/jira/browse/ITK-2939),
itkSymmetricEigenAnalysis returns eigenvalues with the wrong sign when
OrderByMagnitude is set to true.

The first part of this patch is to incorporate a test that
demonstrates this bug.  The second part of the patch will provide the
fix for the bug.

Change-Id: Iddc662a6181fc6b5ded1df83a33f50a4275b9eac