STYLE: Remove dead code ITK_TEMPLATE_EXPLICIT
The ITK_TEMPLATE_EXPLICIT code was a failed
attempt in ITKv3 to make faster compilations
with automatic explicit tempate creation.  The
implementation did not produce proper external
templates and did not result in any compilation
benefits.  This code is being removed because
it no longer can be used by ITKv4 (there are no
build options to include these bits of code).

Change-Id: I220e812c0df8fe32c4e768f86d9da403df63c340