BUG: GDCMImageIO ignores ImagePositionPatient
GDCMImage I/O fails to retain a user provided Image Position Patient
(0x0020, 0x0032). Prior to ITKv4, GDCMImageIO would keep the Image
Posiiton Patient tag if it was specified by the user in ITK's meta
data dictionary. This patch restores that behaviour. A new test
verifies that when provided, Image Position Patient is stored in the
dicom header. If Image Position Patient is not in the meta dictionary
for the image, the IPP is created from the image's origin.

Also, GDCM2 will not write IPP unless the modality is one of CT, MR or
RT. The test sets the modality to CT.

The patch is based on a fix suggested by Alberto Rey (bertorey at
gmail.com) on April 15, 2011.

Change-Id: Ic833627254421929ff54e3bf7004ba5b46347492