STYLE: Made use of .hxx guards consistent
Occasionally there is a need to manually perform explicit instantiation for a project.
A .cxx file is created with the template instantiations where the .hxx files are #include'd,
and add_definitions(-DITK_MANUAL_INSTANTIATION -DITK_TEMPLATE_TXX) is added.
This is very help for identifying which types are used in an algorithm.

This patch is to consistently use only the ITK_MANUAL_INSTANTIATION instead
of an mix of ITK_MANUAL_INSTANTIATION and ITK_TEMPLATE_TXX that are performing
essentially the same task.

Change-Id: I4be7b0ff95e5ced5e1c93b1ccd00d74793ccba02