BUG: Update uncrustify configuration with space before access spec
Privously there was no blank line before public, protected or private
after uncrustification.

Doxygen automatically groups consecutive lines, without spaces. When
one line is an access specifier, then all lines are treated as with
that level of access. The has resulted in numerous methods being
marked as protected or private in the doxygen.

Change-Id: I16a773fb8fd7d68231bc2e2aed95ec9e51ea5fea