COMP: Fix the wrapping for some modules
I have found out that some primarily the convolution and deconvolution
filters have restrictions in dimensional applicability. Additionally the
VtkGlue classes have concept checking which allow only dimensions 1,2 and
3.
A TypedefMacros.cmake bug was fixed to support giving a dimensions list as in
the convolution *.wrap.
Also /bigobj is necessary for proper compilation in VS2005 and higher and
was fixed in ITKSetStandardCompilerFlags.cmake.

VtkGlue wrapping needs vtkPythonCore to link to.

Change-Id: I30912790e76c1de2aa6475b52dd4fae34dbb8573