BUG: Origin was lost reading HDF5 format images.
This fixes this bug report
https://itk.icts.uiowa.edu/jira/browse/ITK-2973

Two problems:

1. The origin was being set before the # of
dimensions, and ImageIOBase::SetNumberOfDimensions()
resizes the origin, spacing etc, and overwrites the
origin member variable.

2. The regression test neglected to set a non-zero
origin in the created test image, so the test would
always succeed, because the origin of the test image
and the image read from disk were both zeroes.

Change-Id: I95e9bc129627307c6d44b23d7e7cf53fe11bbe76