BUG: Correcting valgrind error.
The time varying registration methods were failing on
the valgrind tests and this patch is meant to fix those
failings.  The culprit was a Set macro in both methods.
In order to create the displacement field, one has to
integrate the velocity field.  We do this at discrete
time points using the call

transform->SetUpperTimeBound( x );
transform->SetLowerTimeBound( y );
transform->IntegrateVelocityField();

where the SetUpper/LowerBound employs a SetClampMacro()
with the limits of integration being 0 and 1.

However, the problem was when the upper bound was 1.0.
I'm guessing there's a numerical issue in that 1.0 doesn't
fall in the range dictated by the tests.

Regardless, the tests now pass on my machine (MacOS X 8)
using valgrind 3.8.1.

Change-Id: Idc8bfa9c621f7ee2657de88d9a73383e07669615