BUG: Make LaplRecurGauss filter set the number of threads on the subfilters
The LaplacianRecursiveGaussian filter uses the RecusiveGaussian, Derivative, BinaryFunctor, and CastImage filters, all of which are multithreaded. However, the LaplacianRecursiveGaussian filter does not set the number of threads on these filters, resulting in counter-intuitive behavior when SetNumberOfThreads() is called on the LaplacianRecursiveGaussian filter.

This patch calls filter->SetNumberOfThreads(this->GetNumberOfThreads()); on each of the aforementioned filters and also updates the doxygen documentation in the header file to indiciate that LaplacianRecursiveGaussian filter is actually multithreaded.

Change-Id: I1c3013b6c34bb82b91030f4c70bce471e2449e3e