STYLE: Rename itkMultiTransformBase to itkMultiTransform.
In ITK convention, only classes with a naming conflict between the
base and non-base version have "Base" in their name.  For example,
Image and ImageBase.  Most ITK classes that are base classes do not
have base in their name.  Classes that are not base classes at the
time they are written may often become one in the future.

Change-Id: I9ca801e94d8779b5568b57372b6ded418162b7f4