COMP: Fix link error with visibility-inlines-hidden
A virtual inline function, in HDF4::DataType was causing link errors
when compiled with visibility-inlines-hidden. The method has been
moved to the cpp file.

The follow is the error addressed:

/usr/bin/ld: CMakeFiles/ITKIOHDF5.dir/itkHDF5ImageIO.cxx.o: relocation
R_X86_64_PC32 against undefined symbol `H5::DataType::fromClass()
const' can not be used when making a shared object; recompile with
-fPIC