BUG: Denoising 3D patch weights incorrect
Fixes 3D and greater patch weight computations to be correct.  Also,
only resample the patch weights based on relative anisotropy.  Thus
images with spacing 1,1,1 or 0.5,0.5,0.5 should have the same patch
weights.  While an image with spacing 1,1,0.5 would have more weights
in the z direction so that the weights cover the same physical space
in all 3 directions.  Finally, if the center weight is not 1,
normalize it so that it will be 1.

Change-Id: Ib5cf04a32269ede3eade4ae780e85dfe15fcd4cc