COMP: Remove unnecessary BORLAND code
A universally applied work around for the BORLAND compiler
was causing a compiler warning on Intel compiler where it was
validly detecting multiple declarations of the same forward
declared class across different scoping rules.

This patch removes the work around for the borland compiler,
simplifies the code, and removes a compiler warning when
extensive warnings are used with the Intel compiler.

Change-Id: Ic6ebe6b9649c5b998c8f8bed76a29b907abe8447