ENH:  Adding masking capabilities with confidence mask option.
This is the first of two commits aimed at improving the B-spline
image registration although this commit is an improvement in
and of itself.  This filter now has the option of specifying
a confidence mask.  This allows the user to specify the relative
weighting of the different points in the input displacement
field based on the values specified in the confidence mask
image.

Change-Id: Ic266dc891d2e2a3b61d82667549064239f0353bb