COMP: Fix TIFF on big endian systems
This is similar to the fix I submitted April 24, 2012 (367953).
Basically on big endian systems you need to check if
WORDS_BIGENDIAN is defined, not its value.

This fix was overwritten when libtiff 4.0.3 was pulled in. Also,
libtiff 4.0.3 contained a second instance of this error.

The original full description is here:
     http://lists.debian.org/debian-med/2012/04/msg00070.html