BUG: Do not undefine _WIN32_WINNT
Instead, check to see if the includer has already defined it,
and if so, leave the value alone.

Only if the includer did not define it beforehand should a value
be defined here.

With an undef in there, it may mess up files included afterwards
which expect the value set by the including project, not a different
(and forced!) value from an ITK header.

Change-Id: I3ae8e44c8d0573f6fc4f0f012941a3d309f51941