COMP: Fix robustness of DCMTK internal builds
When Module_ITKIODCMTK is OFF, then the internal DCMTK
should NOT be build in the third party directory.

When Module_ITKIODCMTK is ON, then only build the libraries
because the DCMTK BUILD_APPS:BOOL=ON would unnecessarily
build DCMTK applications that are not needed by ITK.

Use the http rather than https in order to increase the
ability of downloads over firewalled environments.

Change-Id: I1293853bcf6ce7b7e11ef3fbdbf7098ec878d1e6