BUG: Fix "git gerrit-push" for content links with spaces
The sed-to-awk conversion by commit 180a5476 (Fix "git gerrit-push"
content link recognition, 2011-06-13) increased portability but also
broke recognition of content links with spaces in the path.  This is
because the awk "$6" expression matches only the file name up to the
first space.

Remove the string processing and use the bash 'read' command to parse
the diff-tree output lines.  It parses tokens on whitespace but the last
token automatically consumes the rest of the line and preserves spaces.

Change-Id: Ia21258133d0f0a71627607e042b8b8274e9f8611