COMP: Fix all valid cppcheck warnings in "Module/Registration" directory
The cppcheck is run inside the ITK/Module/Core directory and all the
valid cppcheck issues are fixed.

Some of the cppcheck warnings are related to the functions that are not
used, but they may be used later by users. Such a warning is an example of
invalid warnings.

The titles of major fixed issues are listed below:

- prefix ++/-- operators are preferred for non-primitive types.
- initialization to assignment is preferred.
- if possible, the scope of variables are reduced.
- initialization of variable in initialization list is preferred than
assigning variable in constructor body.

All the ITK tests are passed after the above changes.

Change-Id: If152f4b061f5e99ef39af9aa2685f63de5630344