COMP: Fix all valid cppcheck warnings in "Module/Filtering" directory
cppcheck is run inside the ITK/Module/Filtering directory and all the
valid cppcheck issues are fixed.

Some of the cppcheck warnings are related to the functions that are not
used, but they may be used later by users. Such a warning is an example
of an invalid warning.

Some warnings are related to the intentionally created issues that are
provided in the test files. Such warnings are another example of what I call
invalid warnings.

The titles of major fixed issues are listed below:

- if possible, variables are passed by reference.
- prefix ++/-- operators are preferred for non-primitive types.
- initialization to assignment is preferred.
- if possible, the scope of variables are reduced.

All the ITK tests are passed after the above changes.

Change-Id: I2ccc197141fdaa1cc4b70bfab20d5c87a5cf72ca