BUG: CSVFileReader was reading all values as NaN
The ConvertStringToValueType method of CSVFileReaderBase.h was
previously not working correctly.  The operations on the istringstream
resulted in all calls returning NaN even if valid values were passed
in.  This means that every value read from a CSV file was converted to
NaN!  This was not caught by the itkCSVArray2DFileReaderWriterTest
because it was comparing the file that was read to the expected value
without considering NaN values.  This patch corrects the error in
CSVFileReaderBase.h. It also corrects the test so that it will catch
such errors.  In other words, running this updated test on the old
CSVFileReaderBase.h will result in a failed test.

Change-Id: Ic5643ff67cc5fdfb2c7b408e760c0fe471be9be1