STYLE: Add detail to the vxl update instructions
Make the working directory for each command sequence explicit.  State
the purpose of the script recorded in the upstream snapshot branch
commit messages.  Add ascii-art to show the history shape as it is
constructed.  Add a Change-Id line to the upstream snapshot commits so
that the entire update topic can be reviewed in Gerrit.

Change-Id: I849330fe6bdd5c48f70c2f0a7916328936b3e9de