BUG: Remove itkLoad definition in JPEG2000ImageIOFactory
The JPEG2000 ImageIO in Review declares itkLoad,
making the Review module eligible for ObjectFactory
loading if the path to the ITKReview shared library
ends up in ITK_AUTOLOAD_PATH.
But the module defines other IO, which are not registered
like this, if at all. So there is no reason for JPEG2000
to be specific.

My bug case was even more unexpected : among real
ITK modules, the ITK_AUTOLOAD_PATH dir
contained some shared library linked to ITKReview,
but not the ITKReview shared lib itself.
Still, the ObjectFactory mechanism loaded the ITKReview
shared lib as a module.
The side effect of this is that it triggered :
symbol lookup error: /usr/lib/nvidia-current/libGL.so.1
at exit.

Simply removing the itkLoad definition from the ITKReview code
avoids the issue.

Change-Id: I86fd64b5c09490d85a7082b4cf5173a05d16f8ce