BUG: Only thread 0 should report progress in LabelMapFilters
The progress update observer is frequently used to update GUI
progress. The progress should only be updated by the main thread,
otherwise this common use will segfault with the gui update.

The LabelMapFilter was updated with a manual update to the filters
progress, additionally the abort status is also check.

Additionally, the mutex lock was change to use a holder to follow the
Resource Acquisition Is Initialization (RAII), principle. If an
exception occurs the locked block, the lock will automatically be
released.

Change-Id: I001066f59cc336bc8f54ef7299e2f8a030f32ce8