BUG: ITK_FUTURE_LEGACY_REMOVE must cmake configured
The use of the ITK_FUTURE_LEGACY_REMOVE flag must be consistently used
in all ITK compilation units.  If it is not, then very difficult to
debug problems occur in virtual function table lookups.  This problem was
revleaed when ITK was built without ITK_FUTURE_LEGACY_REMOVE, but external
tools were built with ITK_FUTURE_LEGACY_REMOVE flag used (during a migration
to the preferred itkGetModifiableObjectMacro usage).

By making this an ITK cmake option, we can ensure that the header files
properly propogate the desired API consistently between the ITK
libraries and any external tools that depend on them.

Change-Id: Ica2edcf7a2630e1b783ee7fc34aa9ce607df51d2