ENH: index can be associated with required input
Current ITK filters have indexed input without names. To migrate these
classes to have named inputs, names need to be associated with the
current index to maintains compatibility.

The addition of an optional argument to ProcessObject's
AddRquiredInputName method is used to connect the two different
indexing types for the filter's input.

The SetPrimaryInputName method was re-factored to only call existing
methods and not directly manipulate the internal structures.

Change-Id: Iaa20c88eff270f95ada0ea52fba69ebf366a5f45