BUG: Expose BSplineInterpolateImageFunction test
The BSplineInterpolateImageFunction function does not alway produce
the correct answer when the image direction cosines are not
the trivial identity matrix.  This is described in jira issue
https://issues.itk.org/jira/browse/ITK-2889 reported by
Marius Staring.

In addition, some common code was extracted to the makeRandomImageBsplineInterpolator.h
and non-trival direciton cosines are being used for the itkRandomSource
image to improve the amount of stress test provides.

Change-Id: I05417dbe7c87e9577954679c481ff3c29a9cc291