ENH: Added m_RequiredFractionOfOverlappingPixels to MaskedFFTNCC
Added an argument to the Masked FFT NCC algorithm that enables the user
to specify the m_RequiredFractionOfOverlappingPixels.  This provides an
intuitive way for the user to remove noise pixels on the border of the
correlation image.
It is also robust to changing image sizes because it enables
calculation of the numberOfRequiredOverlappingPixels relative to the
maximumNumberOfOverlappingPixels, which is calculated by the filter.
For example, the user can generally set the argument to 0.1 for most
applications since this will remove the pixels that overlap by 10% or
less of the maximum overlap.
This is also particularly important for odd-shaped masks, where it
becomes difficult for the user to know beforehand how many pixels will
overlap during the correlation process.

The filter header file contains detailed description of the new argument.

Updated tests are provided that use the new
m_RequiredFractionOfOverlappingPixels argument.

Change-Id: Ifbf4ebcc43d49e59040b17140c388199449e01a8