COMP: Remove copy_n wrapper for standard std::copy
This patch address VS2010 not compiling from error in
itkStdAlgorithm.h.

While the interface of std::copy_n is close to how we are using it,
the maintenance and burden of creating C++x11 wrappers should be
avoided. So the usage of itk::algorithm::copy_n has been replaced with
std::copy.

Change-Id: I4cb2aefc759c6dabfea5f455b6f93a0f8a798921