ENH: Simplify template parameters in itk::PointsLocator
All current template parameters could be directly deduced from the
PointsContainer template argument. It simplifies the usage of this class in a
general context.

The test has been modified in order to test with different PointsContainer
(itk::VectorContainer<> and itk::MapContainer<>), the 2 main containers used in
itk::PointSet<> and its derivative (itk::Mesh<> and itk::QuadEdgeMesh<>).

Change-Id: I7e9313250355196e6f7b2dbcbb8f796e06a543f0