PERF: Change default index i/o prefix to just "_"
Renamed default index i/os' prefix from "IndexedDataObject" to just "_"
followed by the index number.

There are two motivation for this: 1) The prior approach required each
comparison done in the named map to compare 17 chars to get to the
number, which is inefficient. 2) With the introduction on the global
variable of these name, shortening the name reduces the ( by a small
amount ) the foot print of the variable.

Change-Id: Ifb9e2fb627f11e001f58090af88b71debb421c98