ENH: Add Remote Module for r/w of MGH files
The MGH file format is prevelant in many
neuroimaging environments and it would be
convenient if ITK were able to do at least
basic manipulations of these image file
formats.

This patch is intended to develop a solution
including MGHIO as a Remote Module.

The Remote Module has more extensive testing than the original.

Change-Id: I8d3e3d62062c2236e8aca5760551348c5eaf7494