DOC: Fix line length issues for the SoftwareGuide
Printing code in the latex generated SoftwareGuide requires that the code be
under approximately 80 characters in order to fit in the allocated space.  This
patch set reformats the code so that when printed in the SoftwareGuide it does
not overrun the column boundaries.

Additionally, the code was made more compact so that it fits more tightly into
the prose.  The small bits of code in the book can have less white space than
in the complete source file and still be easy to comprehend.

Change-Id: I472f4d643d39f809fc6afebb5bc2c917dc9834eb