COMP: Fix scanf conversion specification for id types
In a reader just used for a test, scanf was used with a %ld ( signed
long ) conversion specification, while the argument was an unsigned
identifier type. This id type could be 32 or 63 bits, long or long
long. There is currently no macro in ITK to assist with the conversion
specification similar to PRIu64 from inttypes.h. So at a trade off of
not supporting 64-bit integers on windows ( for just a test), the
unsigned long type is used with the consistent %ul conversion
specifier.

Change-Id: I3a24ca1e0fa655a1ac33f957b58e4aaa139868a2