BUG: Wrong classpath separator on Windows
Updated the SCIFIO remote commit that is checked out. This new commit
contains logic to select the correct path separator on Windows and
OSX/Linux.

Previously this was hard-coded to the OSX path separator, which would
cause the itk-bridge-pipes jar to not be found.

Also includes cquammen's fixes for the ImageInfo test and reading
newlines on Windows.

Change-Id: I4c4a864443fedbe6439c18e8a18469c8560a9541