COMP: Do not try to include float.h for MinGW.
float.h is needed for FloatingPointExceptions.  ITK can fail
to compile with MinGW, because there are two float.h files on Windows
in this case.  If it happens that the MinGW one is found first,
a failure will occur because a

  #include_next <float.h>

is not in the GCC version.  While this problem has been reported by others to
upstream GCC many times in the past, it has not been addressed.  To get around
this, we avoid trying to include float.h with MinGW.

Change-Id: I2e8022a5b4c67a6fbef930b5ceda403bcb31c97f