BUG: JPEG2000 bool type in C is now unsigned char
JPEG2000 is a compiled C library interfaced to a C++ library. When
compiling the C library the "bool" type is defined as a int yet in C++
it's a intrinsic bool type, which is 1 byte. While it's not guaranteed
that a bool is the same size, alignment etc as a unsigned
char. In practice its working better until the openjpeg library can be
updated.

Change-Id: I3780812dd60fc23873e5853b35df0f3e3ae5f9dc