BUG: LabelGeometry improvements
This patch corrects the eccentricity, elongation, and orientation
features to operate on the largest eigenvector rather than eigenvector
1. This is the same thing for 2D but different in 3D and higher.

The orientation is also adjusted to always fall between 0 and pi
rather than -pi and pi.  This is possible since the major axis is
symmetric about the origin.  This also makes the orientation values
easier to read and more consistent.

A memory leak in the test was corrected by deleting the matrixPointer
allocated with new.

Change-Id: I12d0d952fe04df43edd118cffb2687f6896870e5