BUG: Corrected the MaskedFFTNCC origin
Corrected the output origin of the MaskedFFTNCC algorithm so that the
fixed image falls directly in the middle of the output correlation map
in physical space.  With this alignment, each point in the output NCC
map overlaps at the location in the fixed image at which the
moving image provides that NCC score when centered at that
location.

Change-Id: I01807fdc49dfd39216da5ec3f0a40861f478d9ce