BUG: GDCMImageIO ignore image orienations from MetaDataDictionary
GCMImageIO can transfer DICOM tags from the MetaDataDictionary
to the DICOM files. However, itkImage attributes of Origin, Spacing,
and Directions are handled specially. Unfortunately, they are also handled
inconsistently.

For Origin, preference is given to DICOM tags in the MetaDataDictionary
over the ITK_Origin entry in the MetaDataDictionary.

For Directions, preference was given to the ITK_ZDirection entry in
the MetaDataDictionary over the DICOM tags in the MetaDataDictionary.

This patch unifies the behavior so the DICOM tags in the MetaDataDictionary
take precedence over ITK_Origin and ITK_ZDirection.

(Note that for Spacing, preference is still given to ITK_Spacing.)

This patch addresses the bug in the 3D Slicer #3254. See
http://na-mic.org/Mantis/view.php?id=3254

This patch set includes a test.

Change-Id: Ibe8cb7dcd7387ba696ba509bd328bca2d77d1831