COMP: Fix warnings in matrix division by scalar.
At compilation time we were observing the following warning in Debian 7
with gcc 4.7:

itkFixedArray.h:197:82: warning: array subscript is above array bounds

A binary disection of the code lead to identifying the division by a scalar of
a VariableSizeMatrix as the culprit.

Expanding the division expression using two nested for loops resolves the
warning.

It is still unclear why a method in the itkVariableSizeMatrix class would
trigger a warning in the itkFixedArray class.

Change-Id: I10a2d4d83cc4d1e5ebc4f9a6cdbcaf134da22c99