BUG: Row/Column assignment into ImageDimensions was wrong
For some reason, I assigned the DICOM Row attribute to
dim[0] and Column attribute to dim[1], which is backwards.
This wasn't caught for a long time because apparently the
overwhelming majority of DICOM slices in the wild are square,
i.e. dim[0] = dim[1].

GDCM always got this right. Now DCMTK is consistent with
GDCM.

Change-Id: I59ba4059b1b0e5dcc55f41308957effb614c44de