STYLE: update SCIFIO module name
As SCIFIO is a remote module, to follow the policies for remote module
naming it should not start with "ITK...".

Thus, the name in the itk_fetch_module() check was changed from
"ITKIOSCIFIO" to "SCIFIO".

The commit hash was also updated to the latest changeset, which makes
the same naming changes to scifio-imageio's itk-module.cmake (so the two
names are still consistent).

Change-Id: Ibaac978152d3d15d26442881c3fb273e81ee6ae9