ENH: Improve internal streaming for SliceBySliceImageFilter
The SliceBySliceImageFilter now only produces the output requested
region, and only request the needed region from the internal
pipeline. The filter still expands the input requested region in the
direction orthogonal to the slice direction.

Check for all input requested region was removes since it's explicitly
set the same and verified to be valid in the PropagateRequestedRegion
method.

Change-Id: Ib10a59e3d83dddbd16f66e1068ea3e041aa23f78