ENH: ResampleImageFilter ReferenceImage only needs to match dimension
ResampleImageFilter can be supplied a ReferenceImage to use to define the
output spacing, origin, and directions. Previously, this ReferenceImage had to be
the same image type as the output.

Since only the meta-data is copied, the ReferenceImage only needs to be the
dimension as the output image.

This patch alters the prototype of SetReferenceImage()/GetReferenceImage()
to take/return pointers to an ImageBase instead of an output image type.

This change is useful, for instance, if you want to resample a labelmap
(image of labels) to match the domain of a VectorImage.

Change-Id: I52f4cb8617e72244a9d49299374dfef5333fcce6