ENH: Rename Utilities/KWStyle/ITK.kws.xml.in to ITK.kws.xml
The file "Utilities/KWStyle/ITK.kws.xml*" is not configured by CMake,
and so should not be named with a ".in" extension. This commit
rectifies this.

This also bumps the version of the SetupForDevelopment script, to ensure
that the pre-commit style checking hooks are configured correctly. Any
local ITK repository that is setup for development will need to re-run
the Utilities/SetupForDevelopment.sh script once again before committing.
Change-Id: I2243cf33e9ad0c1cc8c3dead9695c1ba541becc9