BUG: Enabled negative background values
Prior to this patch, the background value in the output image could
only be non-negative because it was casted to an unsigned integral
type that was used as an index into a vector. Casting negative values
would make the index a large number that fell outside the valid range
of vector indices which to my surprise only sometimes caused a
segmentation fault.

This patch removes the casting and vector indexing, which turned out
not to be used, thereby enabling use of negative background values.

Documentation more clearly describing the non-object pixel intensity
(zero) and the background property was also added.

Change-Id: Ia35096040efd10f4452ba15ab59fe8b7270b1840