COMP: TranformIOBase unused GetTypeNameString function.
This function was declared in an anonymous namespace in the .hxx file, then
used by the derived classes.  This does not work with ITK_MANUAL_INSTANTIATION
and causes and unused-function warning when the object file does not contain
one of the derived classes.

The functions were made into static const inline class members.

Co-authored-by: Kent Williams <norman-k-williams@uiowa.edu>

Change-Id: I403574171caf73e28b9d244755f2cba0dd9e13c2