BUG: Make KWStyleTest output in a uniform and concise format
When invoked without editor-specific formatting (the default on
non-VisualStudio builds), KWStyle produces excessively verbose output,
to the point that test output submitted to CDash gets truncated. Invoking
without the '-v' flag does not solve this issue.

On a Unix system, piping this output of KWStyle to "grep -B1 '^Error'"
would resolve the issue, but there is no cross-platform way to easily
perform this filtering.

Instead, always use GCC-style output for the KWStyle test, and only use
the user-configured output style for the build target. This has the
additional benefit of making the test output uniform across all
configurations.

Change-Id: Ie8c9bb30ea210f4fec0900b7e97c48473e56813c