COMP: HistogramMatchingFilterTest problem with ITKV3_COMPATIBILITY
I recently extended the itkHistogramMatchingImageFilterTest to try
different scalar pixel types.

If ITKV3_COMPATIBILITY is enabled, this test will not compile.

This is because of a type collision between the the
MeasurmentVectorType and IndexType -- if the pixel type is long,
then these types are identical.

This is fixed in ITKV4 by having distinct IncreaseFrequenceOfIndex
and IncreaseFrequencyOfMeasurment methods, instead of overloading
IncreaseFrequency to accomodate the two different attributes.

Bottom line -- you can't use HistogramMatchingFilter on images
with a pixel type of long if ITKV3_COMPATIBILITY is turned on.

This patch just comments out the test that uses long as the pixel
type.

Change-Id: Ib5dc109ae0401691a6ee41855271b7e5503ee5eb