BUG: ReachedTargetPoints not set for OneTarget
When the m_TargetReachedMode was set to OneTarget, the
m_ReachedTargetPoints was always null after running the filter.
This has now been corrected by adding the target point to the
m_ReachedTargetPoints when it is reached.  This was not a problem when
setting m_TargetReachedMode to SomeTargets or AllTargets.
In addition, when m_TargetReachedMode was set to OneTarget or
NoTargets, the program would crash if the m_ReachedTargetPoints
structure was queried: for example when using
filter->GetReachedTargetPoints()->size().  This has also been
corrected by always initializing the m_ReachedTargetPoints structure
in the Initialize method.

Change-Id: I70afb74e337fe97230a76597be23e036e6c21b57