BUG: fixedOrigin corrected in DeformableRegistration examples (ITK-3218)
Several of the DeformableRegistration examples (4, 6, 7, 8, 12, 13, 14, 15)
improperly initialize the fixedOrigin variable, which is used to define the
spline mesh.  An array index is missing from the left hand side of the
expressions, causing all three elements to be set to the same value (most
likely the z origin).  This doesn't matter in many cases (e.g. when the
fixed image is an isotropic volumes centered on the origin), but causes
problems when the volume is not cubic or not centered, as the spline mesh
may not overlap the fixed image volume.  This is a fix for the bug
reported in ITK-3218.

Change-Id: I614b00b067e793b83663b259be08100cca57d8f1